requirements.txt
torch==1.3.1
numpy
tqdm
pyyaml
librosa==0.7.0
numba==0.48.0
tensorflow

